+++ MixViT/lib/test/utils/load_text.py
@@ -34,7 +34,7 @@ def load_text_pandas(path, delimiter, dtype):
         return ground_truth_rect
 
 
-def load_text(path, delimiter=' ', dtype=np.float32, backend='numpy'):
+def load_text(path, delimiter=' ', dtype=np.float64, backend='numpy'):
     if backend == 'numpy':
         return load_text_numpy(path, delimiter, dtype)
     elif backend == 'pandas':
