+++ MixViT/lib/train/dataset/got10k_lmdb.py
@@ -124,7 +124,7 @@ class Got10k_lmdb(BaseVideoDataset):
         bb_anno_file = os.path.join(seq_path, "groundtruth.txt")
         gt_str_list = decode_str(self.root, bb_anno_file).split('\n')[:-1]  # the last line in got10k is empty
         gt_list = [list(map(float, line.split(','))) for line in gt_str_list]
-        gt_arr = np.array(gt_list).astype(np.float32)
+        gt_arr = np.array(gt_list).astype(np.float64)
 
         return torch.tensor(gt_arr)
