+++ tools/convert_sportsmot_to_coco.py
@@ -92,15 +92,15 @@ for split in SPLITS:
         print("{}: {} images".format(seq, num_images))
         if split != "test":
             det_path = os.path.join(seq_path, "det/det.txt")
-            anns = np.loadtxt(ann_path, dtype=np.float32, delimiter=",")
+            anns = np.loadtxt(ann_path, dtype=np.float64, delimiter=",")
             if USE_DET:
-                dets = np.loadtxt(det_path, dtype=np.float32, delimiter=",")
+                dets = np.loadtxt(det_path, dtype=np.float64, delimiter=",")
             if CREATE_SPLITTED_ANN and ("half" in split):
                 anns_out = np.array([
                     anns[i] for i in range(anns.shape[0])
                     if int(anns[i][0]) - 1 >= image_range[0]
                     and int(anns[i][0]) - 1 <= image_range[1]
-                ], np.float32)
+                ], np.float64)
                 anns_out[:, 0] -= image_range[0]
                 gt_out = os.path.join(seq_path, "gt/gt_{}.txt".format(split))
                 fout = open(gt_out, "w")
@@ -116,7 +116,7 @@ for split in SPLITS:
                     dets[i] for i in range(dets.shape[0])
                     if int(dets[i][0]) - 1 >= image_range[0]
                     and int(dets[i][0]) - 1 <= image_range[1]
-                ], np.float32)
+                ], np.float64)
                 dets_out[:, 0] -= image_range[0]
                 det_out = os.path.join(seq_path,
                                        "det/det_{}.txt".format(split))
