+++ yolox/motdt_tracker/reid_model.py
@@ -60,7 +60,7 @@ def load_net(fname, net, prefix='', load_state_dict=False):
                 lr = h5f.attrs['learning_rates']
             else:
                 lr = h5f.attrs.get('lr', -1)
-                lr = np.asarray([lr] if lr > 0 else [], dtype=np.float)
+                lr = np.asarray([lr] if lr > 0 else [], dtype=np.float64)
 
             return epoch, lr
 
@@ -243,14 +243,14 @@ def load_reid_model(ckpt):
 
 
 def im_preprocess(image):
-    image = np.asarray(image, np.float32)
-    image -= np.array([104, 117, 123], dtype=np.float32).reshape(1, 1, -1)
+    image = np.asarray(image, np.float64)
+    image -= np.array([104, 117, 123], dtype=np.float64).reshape(1, 1, -1)
     image = image.transpose((2, 0, 1))
     return image
 
 
 def extract_image_patches(image, bboxes):
-    bboxes = np.round(bboxes).astype(np.int)
+    bboxes = np.round(bboxes).astype(np.int64)
     bboxes = clip_boxes(bboxes, image.shape)
     patches = [image[box[1]:box[3], box[0]:box[2]] for box in bboxes]
     return patches
@@ -261,7 +261,7 @@ def extract_reid_features(reid_model, image, tlbrs):
         return torch.FloatTensor()
 
     patches = extract_image_patches(image, tlbrs)
-    patches = np.asarray([im_preprocess(cv2.resize(p, reid_model.inp_size)) for p in patches], dtype=np.float32)
+    patches = np.asarray([im_preprocess(cv2.resize(p, reid_model.inp_size)) for p in patches], dtype=np.float64)
 
     with torch.no_grad():
         im_var = Variable(torch.from_numpy(patches))
